TV Spielfilm fetcher: parse start/stop times from progress bar
Parsing the start/stop times from the time column caused several issues, especially when transitioning from one day/year to the next.

For example, a program which starts on one day and ends on the next required special handling to ensure that the stop time is really on the next day. Similarly, programs were saved for the wrong year around new year.

Use the progress bar instead. It provides the correct timestamps directly.

fixes #20