TV Spielfilm fetcher: update channel only after fetching everything
The DB and GUI update was triggered after fetching the programs of one day. The TV Spielfilm fetcher fetches the program for tomorrow, today and yesterday to ensure that all programs from 00:00 until 23:59 are available. Therefore, fetching one channel could trigger the DB and GUI update up to three times (depending on what was available already).

Trigger the DB and GUI update only once. This should speed things up on slower systems (e.g. if DB access is slow such as on the PinePhone).