Fetcher implementations: use callbacks
The fetcher implementations had to update the database and emit signals correctly. This was error prone and could not be documented properly in the `FetcherImpl` contract.

Add callbacks to the fetch functions and update the database and GUI (i.e. emit the signals) in one central place (`Fetcher`).

As a side effect, this reduces database accesses which should speed things up on slower systems (e.g. if DB access is slow such as on the PinePhone).