Do not set padding for SwipeListItem in group list
fixes
```
qml: Don't override the leftPadding or rightPadding on a SwipeListItem!
This makes it impossible for me to adjust my layout as I need to for various usecases.
I'll try to fix the mistake for you, but you should remove your overrides from your app's code entirely.
If I can't fix the paddings, I'll fall back to a default layout, but it'll be slightly incorrect and lacks
adaptations needed for touch screens and right-to-left languages, among other things.
```