Rework program description update
The program description is fetched lazy when clicking on a program on the table page to avoid a lot of unnecessary downloads.
So far, this triggered a complete reload of the channel.

Add a dedicated signal to reload only a single program.

Also, fix the problem that the program description was not shown unless the overlay was closed and re-opened.