TV Spielfilm fetcher: handle incorrect stop time
It can happen that the stop time is incorrect (i.e. after the next program has already started).

In this case, all following programs of the same day which started before the incorrect stop time
were merged into the program with the wrong data. All programs of following days which started
before the incorrect stop time were not shown at all.

Overwrite the stop time with the start time of the next program.