TV Spielfilm fetcher: fix program stop time before start time
It can happen that the stop time of a program is incorrect (i.e. after the next program has already started).
In this case, TV Spielfilm fetcher overwrites the stop time with the start time of the next program.

As days are fetched from future to past to be able to stop early,
the last program of the next day is the "previous" program of the first program of the current day.
When the stop time was "fixed" in this case, it led to empty entries and misaligned times on the "Favorites" page.

Fix the stop time only if the "previous" program is really previous.

Note: To avoid problems like this, it would generally be better to parse in chronological order.
However, this would be at the cost of performance (stopping early no longer possible).
Additionally, the cases where fixing the stop time is necessary are only rare exceptions.


BUG: 497738

FIXED-IN: 24.12.1