Database clean-up: keep full days
The database clean-up deletes old programs based on the "delete program after" setting. So far, this also deleted programs of that day up to the current time (e.g. all programs until 11:45h yesterday). This is not what the user would expect intuitively and causes problems with !187.

Keep all programs of the day.