Speed up program description update
When using TV Spielfilm fetcher, the program description is fetched only after clicking on a program to avoid fetching a lot of descriptions nobody ever reads.

This was slow because it updated the database with the description, reset the complete `ProgramsModel` and re-loaded the program (incl. description) from the database.

Emit a signal with the description and update the model and database (for persistence) in parallel.


BUG: 497954

FIXED-IN: 24.12.1