Improve page management
The "Favorites" page existed always to avoid its (expensive) re-creation when switching to a different page (see !136).
However, this caused problems e.g. when selecting favorites: Every time a favorite was added/removed, the `ChannelsModel` of the "Favorites" page was updated. With a certain amount of favorites, the UI became unresponsive.

Delete the "Favorites" page when switching to the "Select Favorites" or "Sort Favorites" page (this will likely change the favorites), keep it in the background when opening the "Settings" or "About" page.

Do not allow to "switch" to an already open page in the global drawer.