Improve handling of gaps/overlaps in the programs
It can happen that the program data contains errors such that there are gaps between programs or programs overlap. If this happens, the "Favorites" page breaks.

This was handled by adjusting the start time of a program to the stop time of the previous program. However, if the stop time of the previous program is completely off (e.g. correct time but on the next day), the problem still persists.

Enhance the logic to (hopefully) be more robust against these kinds of errors. Follow the logic that it's better to start too early than missing something.

BUG: 502258

FIXED-IN: 25.04